# campus-recruitment-plan/校园招聘接口文档包/OpenAPI接口草案.yaml
openapi: 3.1.0
info:
  title: 校园智能招聘与面试预约平台 API
  version: 1.0.0
  description: 面向学生、企业 HR、管理员的校园招聘与面试预约平台 OpenAPI 草案。
servers:
- url: http://localhost:8080/api
  description: 本地开发环境
- url: https://api.example.com/api
  description: 生产环境占位
...
security:
- bearerAuth: []
paths:
...
  /company/profile:
    get:
      tags:
      - Company
      summary: 获取企业资料
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCompanyProfile'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
    put:
      tags:
      - Company
      summary: 保存企业资料
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCompanyProfile'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyProfileSaveRequest'
  /company/certification:
    post:
      tags:
      - Company
      summary: 提交企业认证
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCompanyProfile'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCertificationRequest'
  /company/certification/status:
    get:
      tags:
      - Company
      summary: 获取企业认证状态
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCompanyCertificationStatus'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
  /files/upload:
    post:
      tags:
      - File
      summary: 上传文件
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              - bizType
              properties:
                file:
                  type: string
                  format: binary
                  description: 文件
                bizType:
                  $ref: '#/components/schemas/FileBizType'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseFileUploadResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /files/{fileId}/download:
    get:
      tags:
      - File
      summary: 下载文件
      security:
      - bearerAuth: []
      parameters:
      - name: fileId
        in: path
        required: true
        description: 文件 ID
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: 文件流
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /company/interview/slots:
    get:
      tags:
      - Interview
      summary: 企业面试场次列表
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInterviewSlotPage'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: jobId
        in: query
        required: false
        description: 岗位 ID
        schema:
          type: integer
          format: int32
      - name: status
        in: query
        required: false
        description: 场次状态
        schema:
          type: string
      - *id001
      - *id002
    post:
      tags:
      - Interview
      summary: 创建面试场次
      responses:
        '201':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInterviewSlot'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterviewSlotCreateRequest'
...
  /company/interview/slots/{id}/close:
    put:
      tags:
      - Interview
      summary: 关闭面试场次
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 场次 ID
        schema:
          type: integer
          format: int64
  /company/interview/slots/{id}/bookings:
    get:
      tags:
      - Interview
      summary: 面试场次预约名单
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInterviewBookingPage'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 场次 ID
        schema:
          type: integer
          format: int64
      - *id001
      - *id002
  /interview/slots:
    get:
      tags:
      - Interview
      summary: 查询可预约面试场次
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInterviewSlotList'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: jobId
        in: query
        required: true
        description: 岗位 ID
        schema:
          type: integer
          format: int32
      - name: applicationId
        in: query
        required: true
        description: 投递 ID
        schema:
          type: integer
          format: int32
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CompanyAuditStatus:
      type: string
      enum:
      - UNVERIFIED
      - PENDING
      - APPROVED
      - REJECTED
...
    InterviewSlotStatus:
      type: string
      enum:
      - OPEN
      - FULL
      - CLOSED
      - EXPIRED
    InterviewBookingStatus:
      type: string
      enum:
      - BOOKED
      - CANCELED
      - FINISHED
...
    FileBizType:
      type: string
      enum:
      - RESUME
      - AVATAR
      - LICENSE
      - JOB_ATTACHMENT
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 30002
        message:
          type: string
          example: 重复投递
        data:
          nullable: true
...
    CompanyProfileSaveRequest:
      type: object
      required:
      - companyName
      - city
      - contactName
      - contactPhone
      properties:
        companyName:
          type: string
          example: XX科技有限公司
        industry:
          type: string
          example: 互联网
        scale:
          type: string
          example: 100-499人
        city:
          type: string
          example: 北京
        address:
          type: string
          example: 北京市海淀区
        contactName:
          type: string
          example: 王 HR
        contactPhone:
          type: string
          example: '13800000000'
    CompanyCertificationRequest:
      type: object
      required:
      - companyName
      - city
      - contactName
      - contactPhone
      - licenseFileId
      properties:
        companyName:
          type: string
        industry:
          type: string
        scale:
          type: string
        city:
          type: string
        address:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        licenseFileId:
          type: integer
          format: int64
    CompanyProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        companyName:
          type: string
        industry:
          type: string
        scale:
          type: string
        city:
          type: string
        address:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        licenseFileId:
          type: integer
          format: int64
        auditStatus:
          $ref: '#/components/schemas/CompanyAuditStatus'
        auditReason:
          type: string
    CompanyCertificationStatus:
      type: object
      properties:
        auditStatus:
          $ref: '#/components/schemas/CompanyAuditStatus'
        auditReason:
          type: string
        submitTime:
          type: string
          format: date-time
    FileUploadResponse:
      type: object
      properties:
        fileId:
          type: integer
          format: int64
        originalName:
          type: string
        fileSize:
          type: integer
          format: int64
        fileExt:
          type: string
        accessUrl:
          type: string
...
    InterviewSlotCreateRequest:
      type: object
      required:
      - jobId
      - title
      - startTime
      - endTime
      - capacity
      properties:
        jobId:
          type: integer
          format: int64
        title:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        capacity:
          type: integer
        interviewType:
          type: string
          enum:
          - ONLINE
          - OFFLINE
          default: ONLINE
        location:
          type: string
        description:
          type: string
...
    InterviewSlot:
      type: object
      properties:
        id:
          type: integer
          format: int64
        companyId:
          type: integer
          format: int64
        jobId:
          type: integer
          format: int64
        jobTitle:
          type: string
        title:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        capacity:
          type: integer
        remainCount:
          type: integer
        interviewType:
          type: string
        location:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/InterviewSlotStatus'
...
    InterviewBooking:
      type: object
      properties:
        id:
          type: integer
          format: int64
        slotId:
          type: integer
          format: int64
        applicationId:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        studentName:
          type: string
        companyId:
          type: integer
          format: int64
        companyName:
          type: string
        jobId:
          type: integer
          format: int64
        jobTitle:
          type: string
        status:
          $ref: '#/components/schemas/InterviewBookingStatus'
        bookingTime:
          type: string
          format: date-time
...
    InterviewSlotPage:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/InterviewSlot'
        total:
          type: integer
          format: int64
          example: 100
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
    InterviewBookingPage:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/InterviewBooking'
        total:
          type: integer
          format: int64
          example: 100
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
...
    ApiResponseCompanyProfile:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/CompanyProfile'
    ApiResponseCompanyCertificationStatus:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/CompanyCertificationStatus'
    ApiResponseFileUploadResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/FileUploadResponse'
...
    ApiResponseInterviewSlot:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/InterviewSlot'
    ApiResponseInterviewSlotList:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/InterviewSlotList'
    ApiResponseInterviewSlotPage:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/InterviewSlotPage'
...
    ApiResponseInterviewBookingPage:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/InterviewBookingPage'
...
    InterviewSlotList:
      type: array
      items:
        $ref: '#/components/schemas/InterviewSlot'